integration tests: refactors to describe table
For testing multiple MySQL versions

Simplifies test setup using BeforeAll to move away from the
SynchronizedBeforeSuite - looks like it was a relic of the original
ginkgo version 1

Updates the integration tests to assert on both 8.0 and 8.4

[TNZ-48012](https://vmw-jira.broadcom.net/browse/TNZ-48012)
